requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.3.0
bcrypt==3.1.7
blinker==1.4
boto==2.49.0
boto3==1.16.59
botocore==1.19.59
Brlapi==0.7.0
bubbling-firebase-authentication==0.2.2
CacheControl==0.12.6
cached-property==1.5.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
chardet==4.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-dotenv==1.4.2
django-split-settings==1.0.1
django-storages==1.11.1
djangorestframework==3.11.1
docker==4.1.0
docker-compose==1.25.0
docker-composer==0.8.5
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.17.1
duplicity==0.8.12.0
ecdsa==0.14.1
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
firebase-admin==4.4.0
future==0.18.2
google-api-core==1.26.1
google-api-python-client==1.12.5
google-auth==1.27.1
google-auth-httplib2==0.0.4
google-cloud==0.34.0
google-cloud-core==1.6.0
google-cloud-datastore==2.1.0
google-cloud-firestore==2.0.1
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
gunicorn==20.0.4
httplib2==0.18.1
identify==1.5.6
idna==2.10
importlib-metadata==1.5.0
inflection==0.5.1
iniconfig==1.0.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libcst==0.3.17
lockfile==0.12.2
loguru==0.5.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.0
mypy-extensions==0.4.3
netifaces==0.10.4
nodeenv==1.5.0
oauthlib==3.1.0
olefile==0.46
packaging==20.9
paramiko==2.6.0
pbr==5.5.1
pexpect==4.6.0
Pillow==8.0.0
pluggy==0.13.1
pre-commit==2.7.1
proto-plus==1.17.0
protobuf==3.15.6
psycopg2-binary==2.8.6
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycountry==20.7.3
pycparser==2.20
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
pytest==6.1.1
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-dotenv==0.15.0
python-jose==3.2.0
python-keycloak==0.22.0
pytoml==0.1.21
pytz==2021.1
pyxdg==0.26
PyYAML==5.4.1
reportlab==3.5.34
requests==2.25.1
requests-unixsocket==0.2.0
retrying==1.3.3
rsa==4.7.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
SecretStorage==2.3.1
sib-api-v3-sdk==7.2.0
simplejson==3.16.0
six==1.15.0
soft-delete==0.2
sqlparse==0.3.1
systemd-python==234
testresources==2.0.1
texttable==1.6.2
toml==0.10.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.3
usb-creator==0.3.7
virtualenv==20.1.0
wadllib==1.3.3
webencodings==0.5.1
websocket-client==0.53.0
whitenoise==5.2.0
xkit==0.0.0
zipp==1.0.0
